Trait alloc_wg::alloc::BuildAllocRef

pub trait BuildAllocRef: Sized {
    type Ref: DeallocRef<BuildAlloc = Self>;
#[must_use]
    unsafe fn build_alloc_ref(
        &mut self, 
        ptr: NonNull<u8>, 
        layout: Option<NonZeroLayout>
    ) -> Self::Ref;
}

Associated Types

type Ref: DeallocRef<BuildAlloc = Self>

Loading content...
Required methods

#[must_use] unsafe fn build_alloc_ref(
&mut self,
ptr: NonNull<u8>,
layout: Option<NonZeroLayout>
) -> Self::Ref

Safety

ptr must denote a block of memory currently allocated via this allocator

layout must fit that block of memory

the alignment of the layout must match the alignment used to allocate that block of memory

Loading content...