Trait alloc_wg::iter::IteratorExt

pub trait IteratorExt: Iterator + Sized {
#[must_use = "if you really need to exhaust the iterator, consider `.for_each(drop)` instead"]
    fn collect_in<T: FromIteratorIn<Self::Item, A>, A: AllocRef>(
        self, 
        allocator: A
    ) -> T
    where
        A: Abort,
    { ... }

#[must_use = "if you really need to exhaust the iterator, consider `.for_each(drop)` instead"]
    fn try_collect_in<T: FromIteratorIn<Self::Item, A>, A: AllocRef>(
        self, 
        allocator: A
    ) -> Result<T, CollectionAllocErr<A>> { ... }
}

Provided methods

#[must_use = "if you really need to exhaust the iterator, consider `.for_each(drop)` instead"] fn collect_in<T: FromIteratorIn<Self::Item, A>, A: AllocRef>(
self,
allocator: A
) -> T where
A: Abort,

#[must_use = "if you really need to exhaust the iterator, consider `.for_each(drop)` instead"] fn try_collect_in<T: FromIteratorIn<Self::Item, A>, A: AllocRef>(
self,
allocator: A
) -> Result<T, CollectionAllocErr<A>>

Loading content...